refactor(core): split validation rules and brands
Separate runtime definition rules from compile-time FIX_* brand projections, rename rules after their brands, and enforce parity with a source-level test and documented runtime-only allowlist.

Closes #215
