+++ .changeset/split-validation-rules-brands.md
@@ -1,0 +1,5 @@
+---
+"@crustjs/core": patch
+---
+
+Split runtime validation rules from compile-time `FIX_*` brands and enforce their parity mechanically.
