+++ apps/docs/content/docs/guide/error-handling.mdx
@@ -35,7 +35,7 @@ Crust checks developer-authored definitions and user-authored argv at separate b
 1. **Definition normalization** turns raw flags, arguments, aliases, Contexts, commands, and Extension contributions into trusted Command Nodes. Invalid definitions throw `DEFINITION` while the builder call or command materialization runs. `snapshot()` materializes every Extension command, so a broken nested definition fails even when no user invokes that command.
 2. **Invocation parsing** accepts argv against an already-normalized Command Node. Invalid tokens and conversions throw `PARSE`; missing values and Standard Schema issues throw `VALIDATION`.
 
-Definition Errors do not wait for `run()` or argv parsing. Conversely, an invalid user value is never reported as a Definition Error.
+Definition Errors do not wait for `run()` or argv parsing. Conversely, an invalid user value is never reported as a Definition Error. See [Types and Schemas](./types#runtime-rules-and-type-brands) for the runtime-rule ↔ `FIX_*` brand parity table.
 
 `COMMAND_NOT_FOUND` details contain `input`, `available`, `commandPath`, and `parentCommand`, a readonly serializable Command Snapshot.
