+++ apps/docs/content/docs/guide/types.mdx
@@ -33,6 +33,23 @@ Use `parse` for a small synchronous transformation when a full schema is unneces
 
 The Command Action receives the return type. A parser must be synchronous: TypeScript rejects Promise-returning parsers with `FIX_ASYNC_PARSE`, and runtime normalization rejects declared `async` functions at definition or materialization time for plain-JS consumers.
 
+## Runtime rules and type brands
+
+Runtime normalization is the source of truth. For literal-typed TypeScript callers, Crust projects the same rules into `FIX_*` brands so the editor can point at the offending definition before execution. The [error-handling guide](./error-handling) explains when Definition Errors surface.
+
+| Runtime rule       | TypeScript brand        |
+| ------------------ | ----------------------- |
+| `aliasCollision`   | `FIX_ALIAS_COLLISION`   |
+| `noPrefix`         | `FIX_NO_PREFIX`         |
+| `asyncParse`       | `FIX_ASYNC_PARSE`       |
+| `duplicateArg`     | `FIX_DUPLICATE_ARG`     |
+| `variadicPosition` | `FIX_VARIADIC_POSITION` |
+| `aliasShape`       | `FIX_ALIAS_SHAPE`       |
+| `commandCollision` | `FIX_COMMAND_COLLISION` |
+| `duplicateContext` | `FIX_DUPLICATE_CONTEXT` |
+
+Some runtime rules intentionally have no type-level counterpart because widened TypeScript types cannot project them reliably. The parity test's runtime-only allowlist is: `nonEmptyName`, `reservedSpelling` (`__proto__`), `defaultWithinChoices`, `definitionProvenance`, `duplicateExtension`, `missingContextDependency`, `contextFlagPairCollision`, and `contextCycle` (the latter `duplicateExtension` and `contextFlagPairCollision` checks are enforced inline during command materialization rather than as named rule functions). `schemaExclusivity` and `parserType` also need no `FIX_*` brand because the definition unions enforce them structurally.
+
 ## Schema-backed definitions
 
 Crust uses one Standard Schema contract across core, store, and prompts: put the schema on a plain definition's `schema` key. The schema receives the surface's raw value (including `undefined` when an argument or persisted field is missing), and the caller receives its exact output type. Any Standard Schema v1 vendor works, and validation may be synchronous or asynchronous.
