+++ packages/core/src/api/extension.ts
@@ -9,7 +9,7 @@ import type {
 	NamedFlagDef,
 	NamedFlagsRecord,
 } from "../types.ts";
-import type { ValidateNamedFlagDefs } from "../validation/flags.ts";
+import type { ValidateNamedFlagDefs } from "../validation/flags.brands.ts";
 import { normalizeFlag } from "../validation/normalize.ts";
 import type { Awaitable } from "./context.ts";
