+++ packages/core/src/validation/commands.brands.test.ts
@@ -1,12 +1,10 @@
 import { describe, expect, it } from "bun:test";
 
-import { createCommandNode } from "../command/node.ts";
-import {
-	type CommandDefinitionSpellings,
-	validateIncomingAliases,
-	type ValidateCommandConfig,
-	type ValidateCommandDefinitions,
-} from "./commands.ts";
+import type {
+	CommandDefinitionSpellings,
+	ValidateCommandConfig,
+	ValidateCommandDefinitions,
+} from "./commands.brands.ts";
 
 type Equal<A, B> =
 	(<T>() => T extends A ? 1 : 2) extends <T>() => T extends B ? 1 : 2 ? true : false;
@@ -95,64 +93,5 @@ describe("compile-time command validation", () => {
 		type _check = Expect<Equal<keyof Widened, never>>;
 
 		expect(true).toBe(true);
-	});
-});
-
-describe("validateIncomingAliases", () => {
-	function sibling(name: string, aliases?: readonly string[]) {
-		const node = createCommandNode(name);
-		if (aliases) node.meta.aliases = aliases;
-		return node;
-	}
-
-	it("accepts non-colliding command aliases", () => {
-		expect(() =>
-			validateIncomingAliases(
-				{ canonicalName: "version", aliases: ["v"] },
-				{ issue: sibling("issue", ["issues", "i"]) },
-				"version",
-			),
-		).not.toThrow();
-	});
-
-	it("rejects an alias colliding with a sibling canonical name", () => {
-		expect(() =>
-			validateIncomingAliases(
-				{ canonicalName: "compile", aliases: ["build"] },
-				{ build: sibling("build") },
-				"compile",
-			),
-		).toThrow(/collides with sibling canonical name "build"/);
-	});
-
-	it("rejects aliases colliding across sibling commands", () => {
-		expect(() =>
-			validateIncomingAliases(
-				{ canonicalName: "info", aliases: ["i"] },
-				{ issue: sibling("issue", ["i"]) },
-				"info",
-			),
-		).toThrow(/collides with alias of sibling "issue"/);
-	});
-
-	it("rejects a canonical name colliding with a sibling alias", () => {
-		expect(() =>
-			validateIncomingAliases({ canonicalName: "i" }, { issue: sibling("issue", ["i"]) }, "i"),
-		).toThrow(/canonical name "i" collides with alias of sibling "issue"/);
-	});
-
-	it("rejects duplicate and shape-invalid aliases", () => {
-		expect(() =>
-			validateIncomingAliases({ canonicalName: "issue", aliases: ["i", "i"] }, {}, "issue"),
-		).toThrow(/lists alias "i" more than once/);
-		expect(() =>
-			validateIncomingAliases({ canonicalName: "issue", aliases: ["my issue"] }, {}, "issue"),
-		).toThrow(/must not contain whitespace/);
-	});
-
-	it('rejects aliases beginning with "-"', () => {
-		expect(() =>
-			validateIncomingAliases({ canonicalName: "issue", aliases: ["--issues"] }, {}, "issue"),
-		).toThrow('alias "--issues" must not start with "-" (reserved for flags)');
 	});
 });
