+++ packages/core/src/validation/contexts.brands.ts
@@ -1,5 +1,4 @@
 import type { ContextInstance, ContextMap } from "../api/context.ts";
-import { CrustError } from "../errors.ts";
 import type { DefName, Overlap } from "./shared.ts";
 
 // ────────────────────────────────────────────────────────────────────────────
@@ -75,81 +74,3 @@ export type ValidateContextNames<
 };
 
 // ────────────────────────────────────────────────────────────────────────────
-// Runtime validation
-// ────────────────────────────────────────────────────────────────────────────
-
-/**
- * Validate one incoming Context instance against those already provided on
- * the command path. The type system already rejects both misuses; this
- * catches plain-JS callers.
- */
-export function validateIncomingContext(
-	instance: ContextInstance,
-	existing: readonly ContextInstance[],
-): void {
-	// Catches plain-JS misuse, most commonly passing the factory instead
-	// of an instance (.provide(db) instead of .provide(db())).
-	if ((instance as Partial<ContextInstance> | null)?.kind !== "context") {
-		throw new CrustError(
-			"DEFINITION",
-			"provide() requires Context instances — invoke the factory returned by defineContext() (e.g. .provide(db(options)))",
-			{ subject: "context", reason: "not-a-context" },
-		);
-	}
-	if (existing.some((entry) => entry.name === instance.name)) {
-		throw new CrustError(
-			"DEFINITION",
-			`Context "${instance.name}" is already provided on this command path`,
-			{
-				subject: "context",
-				name: instance.name,
-				reason: "duplicate-context",
-			},
-		);
-	}
-}
-
-/** Order Context instances topologically and reject incomplete or cyclic graphs. */
-export function sortContexts(
-	contexts: readonly ContextInstance[],
-	where: string,
-): ContextInstance[] {
-	const provided = new Set(contexts.map((context) => context.name));
-	for (const context of contexts) {
-		for (const dep of context.requiredCtx) {
-			if (!provided.has(dep)) {
-				throw new CrustError(
-					"DEFINITION",
-					`Context "${context.name}" requires Context "${dep}", which is not provided on ${where}`,
-					{
-						subject: "context",
-						name: context.name,
-						reason: "missing-context-dependency",
-					},
-				);
-			}
-		}
-	}
-
-	const sorted: ContextInstance[] = [];
-	const constructed = new Set<string>();
-	let remaining = [...contexts];
-	while (remaining.length > 0) {
-		const ready = remaining.filter((context) =>
-			context.requiredCtx.every((dep) => constructed.has(dep)),
-		);
-		if (ready.length === 0) {
-			const names = remaining.map((context) => `"${context.name}"`).join(", ");
-			throw new CrustError("DEFINITION", `Contexts ${names} form a dependency cycle on ${where}`, {
-				subject: "context",
-				reason: "context-cycle",
-			});
-		}
-		for (const context of ready) {
-			sorted.push(context);
-			constructed.add(context.name);
-		}
-		remaining = remaining.filter((context) => !constructed.has(context.name));
-	}
-	return sorted;
-}
